# doc/api.yml
openapi: 3.0.3
info:
  title: geo-jobs
  description: Launch Zone Tiling Jobs (ZTJ) and Zone Detection Jobs (ZDJ).
  version: 'latest'
servers:
  - url: 'https://api.birdia.fr'
paths:
...
  '/detectionJobs':
    get:
      tags:
        - MachineDetection
      operationId: getDetectionJobs
      description: List all ZDJ
      parameters:
        - in: query
          name: page
          schema:
            $ref: "#/components/schemas/Page"
        - in: query
          name: pageSize
          schema:
            $ref: "#/components/schemas/PageSize"
      responses:
        200:
          description: All ZDJ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ZoneDetectionJob"
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/recomputedStatus':
    get:
      tags:
        - MachineDetection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: getRecomputedZDJStatus
      description: Get last recomputed ZDJ status while submitting new recomputing again
      responses:
        200:
          description: The recomputed status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/taskStatistics':
    get:
      tags:
        - MachineDetection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: getDetectionTaskStatistics
      description: Get detection task statistics for a given ZDJ
      responses:
        200:
          description: List of detection task statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatistic"
        '500':
          $ref: '#/components/responses/500'
...
  '/detectionJobs/{id}/detectedParcels':
    get:
      tags:
        - MachineDetection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: getZDJParcels
      description: List all detected parcels from a ZTJ
      responses:
        200:
          description: All detected parcels for given ZTJ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DetectedParcel"
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/geojsonsUrl':
    get:
      tags:
        - MachineDetection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: getZDJGeojsonsUrl
      description: |
        Get a secured URL to a ZIP containing the detection results as GeoJSON,
        one GeoJSON per objectType for the whole job.
      responses:
        200:
          description: Information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoJsonsUrl'
        '500':
          $ref: '#/components/responses/500'
  '/detectionJobs/{id}/process':
    post:
      tags:
        - MachineDetection
      parameters:
        - in: path
          name: id
          required: true
          description: ZDJ identifier
          schema:
            type: string
      operationId: processZDJ
      description: Process a ZDJ
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/DetectableObjectConfiguration"
      responses:
        200:
          description: The processed (-ing) ZDJ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneDetectionJob"
        '500':
          $ref: '#/components/responses/500'
  '/usage':
    get:
      tags:
        - Detection
      operationId: getDetectionUsage
      summary: get detection usage for community users
      parameters:
        - in: query
          name: surfaceUnit
          required: false
          schema:
            $ref: '#/components/schemas/DetectionSurfaceUnit'
      responses:
        200:
          description: The statistics of detections usages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetectionUsage'
  '/detections':
    get:
      tags:
        - Detection
      operationId: getDetections
      parameters:
        - in: query
          name: page
          schema:
            $ref: "#/components/schemas/Page"
        - in: query
          name: pageSize
          schema:
            $ref: "#/components/schemas/PageSize"
      responses:
        200:
          description: List of detections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Detection'
  '/detections/{id}':
    get:
      tags:
        - Detection
      operationId: getProcessedDetection
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      responses:
        200:
          description: Detection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
    post:
      tags:
        - Detection
      operationId: processDetection
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDetection'
      responses:
        200:
          description: Detection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/roofer':
    post:
      tags:
        - Detection
      operationId: processRooferDetection
      summary: Process detection for roof
      description: |
        This detection is handled synchronously under 30 seconds and used especially for roof under 100 square metre
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDetection'
      responses:
        200:
          description: Detection result for roof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/roofer/email':
    post:
      tags:
        - Detection
      operationId: sendMailAboutProspect
      summary: Send information about prospect to roofer
      security:
        - ApiKeyAuth: [ ]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Prospect'
      responses:
        200:
          description: Detection result for roof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/geojson':
    post:
      tags:
        - Detection
      operationId: finalizeShapeConfig
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        description: feature from shape
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Detection with updated geojson file uploaded url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/image':
    post:
      tags:
        - Detection
      operationId: configureRooferDetectionImageConfig
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        description: Image from a specific address
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Detection with updated image file uploaded url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/pdf':
    post:
      tags:
        - Detection
      operationId: uploadRooferDetectionPdfResult
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        description: Pdf of the detection result
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Detection with updated pdf file uploaded url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/shape':
    post:
      tags:
        - Detection
      operationId: configureDetectionShapeFile
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Detection with updated shape file uploaded url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
  '/detections/{id}/excel':
    post:
      tags:
        - Detection
      operationId: configureDetectionExcelFile
      parameters:
        - in: path
          name: id
          schema:
            type: string
          description: Detection identifier
          required: true
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: Detection with updated excel file uploaded url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detection'
...
components:
  securitySchemes:
    ReadmeAuth:
      type: apiKey
      name: readme-signature
      in: header
  responses:
...
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    RestException:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/RestException'
      example:
        type: InternalServerException
        message: Unexpected error
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
    GeoServerProperties:
      type: object
      properties:
        geoServerUrl:
          type: string
        geoServerParameter:
          $ref: "#/components/schemas/GeoServerParameter"
...
    DetectionSurfaceUnit:
      type: string
      enum:
        - SQUARE_DEGREE
    DetectionSurfaceValue:
      type: object
      required:
        - value
        - unit
      properties:
        value:
          type: number
          format: double
        unit:
          $ref: '#/components/schemas/DetectionSurfaceUnit'
    DetectionUsage:
      type: object
      required:
        - totalUsedSurface
        - remainingSurface
        - lastDatetimeSurfaceUsage
        - maxAuthorizedSurface
      properties:
        maxAuthorizedSurface:
          $ref: '#/components/schemas/DetectionSurfaceValue'
        totalUsedSurface:
          $ref: '#/components/schemas/DetectionSurfaceValue'
        remainingSurface:
          $ref: '#/components/schemas/DetectionSurfaceValue'
        lastDatetimeSurfaceUsage:
          type: string
          format: date-time
...
    CreateZoneDetectionJob:
      type: object
      properties:
        emailReceiver:
          type: string
        zoneName:
          type: string
        zoneTilingJobId:
          type: string
        type:
          $ref: '#/components/schemas/ZoneDetectionType'
        objectsToDetect:
          type: array
          items:
            $ref: "#/components/schemas/DetectableObjectConfiguration"
    ZoneDetectionJob:
      allOf:
        - $ref: "#/components/schemas/CreateZoneDetectionJob"
      properties:
        id:
          type: string
        creationDatetime:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/Status"
    ZoneDetectionType:
      enum:
        - MACHINE
        - HUMAN
    DetectableObjectConfiguration:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/DetectableObjectType"
        confidence:
          type: number
          description: between 0 and 1
        bucketStorageName:
          type: string
          description: |
            Bucket where are stores images to be detected. 
            If null is provided, the default apps bucket is chosen.
    DetectableObjectType:
      type: string
      enum:
        - TOITURE_REVETEMENT
        - PANNEAU_PHOTOVOLTAIQUE
        - PISCINE
        - PASSAGE_PIETON
        - ARBRE
        - TROTTOIR
        - LINE
        - ESPACE_VERT
        - VOIE_CARROSSABLE
        - MOISISSURE
        - USURE
        - FISSURE_CASSURE
        - OBSTACLE
        - CHEMINEE
        - HUMIDITE
        - RISQUE_FEU
        - PARKING
        - VELUX
        - ESPACE_VERT_PARKING
        - BATI_TUILES
        - BATI_BETON
        - BATI_ARDOISE
        - BATI_AUTRES
...
    DetectedParcel:
      properties:
        id:
          type: string
        detectionJobIb:
          type: string
        parcelId:
          type: string
        creationDatetime:
          type: string
          format: date-time
        detectedTiles:
          type: array
          items:
            $ref: "#/components/schemas/DetectedTile"
        status:
          $ref: "#/components/schemas/Status"
    TaskStatistic:
      type: object
      properties:
        jobId:
          type: string
        jobType:
          $ref: '#/components/schemas/JobType'
        actualJobStatus:
          $ref: '#/components/schemas/Status'
        taskStatusStatistics:
          type: array
          items:
            $ref: '#/components/schemas/TaskStatusStatistic'
        updatedAt:
          type: string
          format: date-time
    TaskStatusStatistic:
      allOf:
        - $ref: '#/components/schemas/StatusStatistic'
    StatusStatistic:
      type: object
      properties:
        progression:
          $ref: '#/components/schemas/ProgressionStatus'
        healthStatistics:
          type: array
          items:
            $ref: '#/components/schemas/HealthStatusStatistic'
...
    TileInfo:
      properties:
        size:
          type: object
          properties:
            width:
              type: integer
            height:
              type: integer
        coordinates:
          type: object
          properties:
            x:
              type: integer
            y:
              type: integer
            z:
              type: integer
    DetectedTile:
      properties:
        tileId:
          type: string
        creationDatetime:
          type: string
          format: date-time
        tileInfo:
          $ref: "#/components/schemas/TileInfo"
        detectedObjects:
          type: array
          items:
            $ref: "#/components/schemas/DetectedObject"
        status:
          $ref: "#/components/schemas/Status"
        bucketPath:
          type: string
    DetectedObject:
      properties:
        detectedObjectType:
          $ref: "#/components/schemas/DetectableObjectType"
        feature:
          $ref: "#/components/schemas/Feature"
        confidence:
          type: number
        detectorVersion:
          type: string
    JobType:
      type: string
      enum:
        - DETECTION
        - TILING
        - PARCEL_DETECTION
...
    Status:
      type: object
      properties:
        progression:
          type: string
          enum:
            - PENDING
            - PROCESSING
            - FINISHED
        health:
          type: string
          enum:
            - SUCCEEDED
            - FAILED
            - UNKNOWN
            - RETRYING
        creationDatetime:
          type: string
          format: date-time

    ###########################################################################################################
    ###  GeoJSON objects. Modified from: https://gist.github.com/bubbobne/fe5f2db65acf039be6a9fd92fc9c7233  ###
    ###########################################################################################################
    GeoServerParameter:
      type: object
      properties:
        service:
          type: string
        request:
          type: string
        layers:
          type: string
        styles:
          type: string
        format:
          type: string
        transparent:
          type: boolean
        version:
          type: string
        width:
          type: integer
        height:
          type: integer
        srs:
          type: string
    Feature:
      type: object
      properties:
        id:
          type: string
        zoom:
          type: integer
        geometry:
          oneOf:
            - $ref: '#/components/schemas/Point'
            - $ref: '#/components/schemas/Polygon'
            - $ref: '#/components/schemas/MultiPolygon'
        properties:
          type: object
          additionalProperties: true
...
    GeoJsonsUrl:
      type: object
      properties:
        url:
          type: string
        status:
          $ref: "#/components/schemas/Status"
...
    DetectionStepName:
      enum:
        - CONFIGURING
        - TILING
        - MACHINE_DETECTION
        - HUMAN_DETECTION
        - GEO_JSON_CONVERSION
    DetectionStepStatistic:
      allOf:
        - $ref: '#/components/schemas/StatusStatistic'
    DetectableObjectModel:
      discriminator:
        propertyName: modelName
      oneOf:
        - $ref: '#/components/schemas/BPToitureModel'
        - $ref: '#/components/schemas/BPLomModel'
        - $ref: '#/components/schemas/BPZanModel'
        - $ref: '#/components/schemas/BPClimatResilienceModel'
        - $ref: '#/components/schemas/BPConformitePluModel'
        - $ref: '#/components/schemas/BPOldModel'
        - $ref: '#/components/schemas/BPTrottoirsModel'
...
    BPLomModel:
      type: object
      properties:
        modelName:
          type: string
          enum: [ 'BP_LOM' ]
        passagePieton:
          type: boolean
          default: true
        voieCarrosable:
          type: boolean
          default: true
        trottoir:
          type: boolean
          default: true
        parking:
          type: boolean
          default: true
...
    CreateDetection:
      type: object
      properties:
        emailReceiver:
          type: string
          description: |
            Email address that will be notified about detection steps evolution
        zoneName:
          type: string
          description: |
            Name that consumers provide to identify the zone
        geoServerProperties:
          $ref: '#/components/schemas/GeoServerProperties'
          description: Optional Configuration properties for the GeoServer.
        detectableObjectModel:
          $ref: '#/components/schemas/DetectableObjectModel'
        geoJsonZone:
          type: array
          items:
            $ref: '#/components/schemas/Feature'
          description: |
            Optional. If so, must provide zone by shape or excel file through its specific endpoint.
    DetectionStep:
      properties:
        name:
          $ref: '#/components/schemas/DetectionStepName'
        status:
          $ref: '#/components/schemas/Status'
        statistics:
          type: array
          items:
            $ref: '#/components/schemas/DetectionStepStatistic'
        updatedAt:
          type: string
          format: date-time
...
    Detection:
      allOf:
        - $ref: '#/components/schemas/CreateDetection'
      properties:
        id:
          type: string
        step:
          $ref: '#/components/schemas/DetectionStep'
        geoJsonUrl:
          type: string
        shapeUrl:
          type: string
        excelUrl:
          type: string
        imageUrl:
          type: string
        pdfUrl:
          type: string
    Prospect:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        address:
          type: string